L1/api/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
numpy
Pillow
python-multipart
